+++ changelog.md
@@ -7,3 +7,9 @@ Initial commit of presets module, creating:
 * 192k mp3
 * 720p mp4
 * 1080p mp4
+
+## 1.1
+
+Add queue module, queue has notifications:
+* COMPLETE/ERROR to SQS
+* All to Cloudwatch logs